Return the license header to Fileformat and Person
License header was auto-deleted in accident by Eclipse due to the fact
that it was located below the package declaration, so Eclipse considered
it a comment on an import. It removed that comment when it removed the
corresponding import.

License headers have been restored above the package declaration so that
this does not happen again.